fix(prometheusrule): fix rule duration unit conversion error
